requirements.txt
Flask>=2.3,<4
reportlab>=4.0,<5
python-docx>=1.1,<2
flask-cors>=4.0,<5
Flask-Session>=0.6,<1
python-dotenv>=1.0,<2
requests>=2.31,<3
